# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API BASE",
    "version": "1.0.0",
    "description": "API de base pour la gestion de l'authentification et des logs.",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Mr NJIFANDA",
      "url": "https://njifanda.com",
      "email": "contact@njifanda.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Base - API (mode: Development)"
    }
  ],
  "paths": {
...
    "/app/user/change-password": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Change password",
        "description": "Change current user password",
        "operationId": "changePassword",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "current_password",
                  "new_password",
                  "new_password_confirm"
                ],
                "properties": {
                  "current_password": {
                    "type": "string",
                    "format": "password",
                    "example": "CurrentPassWORD@2025"
                  },
                  "new_password": {
                    "type": "string",
                    "format": "password",
                    "example": "NewPassWORD@2025",
                    "description": "Must contain: lowercase, uppercase, digit, special char (3-30 chars)"
                  },
                  "new_password_confirm": {
                    "type": "string",
                    "format": "password",
                    "example": "NewPassWORD@2025"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/app/user/login-history": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get login history",
        "description": "Retrieve login history for current user",
        "operationId": "getLoginHistory",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of records to retrieve",
            "required": false,
            "schema": {
              "type": "integer",
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "login_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "ip_address": {
                            "type": "string",
                            "example": "192.168.1.1"
                          },
                          "user_agent": {
                            "type": "string"
                          },
                          "location": {
                            "type": "string",
                            "example": "Paris, France"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/app/user/logout-other-devices": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Logout from other devices",
        "description": "Logout from all other devices except the current one",
        "operationId": "logoutOtherDevices",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/app/user/account": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete account",
        "description": "Delete (deactivate) current user account",
        "operationId": "deleteAccount",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "Current password for confirmation",
                    "example": "PassWORD@2025"
                  },
                  "reason": {
                    "type": "string",
                    "description": "Optional reason for account deletion",
                    "example": "No longer need the service"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/app/user/account/stats": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get account statistics",
        "description": "Retrieve statistics and metrics for current user account",
        "operationId": "getAccountStats",
        "security": [
          {
            "apiKeyAuth": [],
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Account statistics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "example": false
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total_logins": {
                          "type": "integer",
                          "example": 150
                        },
                        "last_login": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "account_age_days": {
                          "type": "integer",
                          "example": 365
                        },
                        "uploads_count": {
                          "type": "integer",
                          "example": 25
                        },
                        "storage_used": {
                          "type": "integer",
                          "example": 52428800
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key required for all endpoints"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authenticated endpoints"
      }
    },
    "responses": {
      "SuccessResponse": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Operation successful"
                },
                "data": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
      "BadRequestError": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Bad request"
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized - Invalid or missing authentication",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Validation failed"
                },
                "errors": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "field": {
                        "type": "string"
                      },
                      "message": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ],
...
}